DRY up: Extract #remove_index override into Module
